% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{blogsville}
\alias{blogsville}
\title{blogsville}
\format{A data frame with eight variables:
\describe{
\item{\code{dates}}{Day of the observation}
\item{\code{uxx}}{Surface Zonal Velocity Predictor variable}
\item{\code{vxx}}{Surface Meridonal Velocity Predictor variable}
\item{\code{zxx}}{Surface Vorticity Predictor variable}
\item{\code{xx500}}{500 hPa Geopotential Height Predictor variable}
\item{\code{humxx}}{Surface Specific Humidity Predictor variable}
\item{\code{tmax}}{Temperature Maximum Predictand variable}
\item{\code{pcrp}}{Precipitation Amount Predictand variable}
}}
\usage{
blogsville
}
\description{
A dataframe containing predictor and predictand variables
from the blogsville example used in the SDSM tool.
}
\keyword{datasets}

