% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sdsm-calibrate.R
\name{calibrate_model}
\alias{calibrate_model}
\title{Build a Linear Regression for Statistical Downscaling}
\usage{
calibrate_model(dataframe, y, model_type = "annual",
  autoregression = "false", process = "unconditional", date_column = NULL)
}
\arguments{
\item{dataframe}{a dataframe object which contains the data
you are fitting the model with}

\item{y}{the repsonse variable. Input must be a character vector}

\item{model_type}{"annual", "monthly", or "seasonal". The default
used is "annual"}

\item{autoregression}{Whether or not to include an autoregressive term
in the model}

\item{process}{Either conditional or unconditional}

\item{date_column}{Column which contains the dates. Should be
of the class Date in R to work, input must be a character!}
}
\value{
A list of either one, four, or twelve linear models, with
the length of the list determined by the model_type parameter.
}
\description{
This function is meant to replicate the calibrate model
section of the SDSM tool. It will date in a dataframe, the response
variable, the dates column, as well as other specific modeling
oprions. As an output, calibrate_model with generate a list
with appropriate number of linear models. The reason it outputs
a list of models, rather than just one, is for consistency when
the "monthly" or "seasonal" option is chosen, which will fit 4 or 12
seperate models to the dataframe.
}

