% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sdsm-functions.R
\name{combine_predictors}
\alias{combine_predictors}
\title{Combines individual predictor files}
\usage{
combine_predictors(file_list, var_names = NULL)
}
\arguments{
\item{file_list}{A list of files path for which t
here are vectors of data we want to combine into one dataframe}

\item{var_names}{An optional character vector which specifies the
names of the columes in the resulting dataframe}
}
\value{
A dataframe of the equal length vectors merged together
}
\description{
This function will take a list of files which has vectors
of equal length, and combine them into one dataframe. It is
ideal when each one of the predictor variables comes in
it's own file.
}

