% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sdsm-functions.R
\name{generate_table}
\alias{generate_table}
\title{Generate diagnostic plots for predictors and predictands.}
\usage{
generate_table(plotname, dataframe, y)
}
\arguments{
\item{plotname}{The filepath in which you want to save t}

\item{dataframe}{a dataframe object which contains a column
of dates, the predictand (response) variable, as well as the
predictor variablesto explore.}

\item{y}{The predictand (response) variable name from the
same dataframe}
}
\value{
A correlation matrix between predictor variables and
two PDF file diagnostic plots.
}
\description{
This function is meant to replicate the screen variables
section of the SDSM tool. It creates two plots, the table
which shows the correlations between predictors in all individual
months and annually, as well as the
}

