% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sdsm-functions.R
\name{generate_weather}
\alias{generate_weather}
\title{Make predictions using calibrated models and a new data frame}
\usage{
generate_weather(models, new_dataframe, uncertainty = "ensemble",
  num_ensembles = 1)
}
\arguments{
\item{models}{A list containing either 1, 4, or 12 models depending
on whether the model is annual, seasonal, or monthly.}

\item{new_dataframe}{A new dataframe in which predictions are to
be made from the calibrated models.}

\item{uncertainty}{Either "ensembles" or "intervals".}

\item{num_ensembles}{If uncertainty = "ensembles", then this specifies
how many ensembles to return. Default is set to one.}
}
\value{
A two column dataframe with dates and predicted weather. The dataframe
is ordered chronologically from the dates column.
}
\description{
This function is meant to be used after running the calibrate_models
function. The idea is that it checks the model type, and then works
through the different subsets of the new dataframe to make predictions,
and returnsa data frame with dates, and predictions, and ensembles
of predictions if that option has been specified.
}

