% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/quality-control.R
\name{quality_control}
\alias{quality_control}
\title{Check to make sure the data-set is appropriate for sdsmR.}
\usage{
quality_control(dataframe)
}
\arguments{
\item{dataframe}{The name you want to save the plots as. IE
if your data comes from the blogsville data-set, you might
want to call this "blogsville". If you want to save the plot
in a different directory, just specify this in this name:
"/home/blogsville" will save the plot's in the home directory.}
}
\value{
Either an error message specifying what went wrong
or print statements verifying all of the components of your
dataframe
}
\description{
This function is meant to replicate the quality control
section of the SDSM tool. It takes in the data-frame you
are hoping to use and verify that it has all the components
needed to work with the other sdsmR functions
}
\examples{
# Check the Blogsville data-frame for errors
quality_control(blogsville)

# Remove the Date column and to return an error
quality_control(blogsville[, 2:5])
}

