% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sdsmR.R
\docType{package}
\name{sdsmR}
\alias{sdsmR}
\alias{sdsmR-package}
\title{sdsmR: A package for performing Statistical Downscaling in
the R language}
\description{
This package provides two forms of common functions.
First, it allows you to perform the exact model
calibration, summary statistics, and prediction you would
normally do in SDSM. Next, it allows you to access NCEP
Re analysis data, which is used to calibrate sdsm models.
}
\section{sdsmR functions}{

\code{\link{generate_table}}
}

