% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sdsm-calibrate.R
\name{summarize_models}
\alias{summarize_models}
\title{Obtain summary statistics for calibrated models}
\usage{
summarize_models(model_list)
}
\arguments{
\item{model_list}{A list build with the calibrate_model function}
}
\value{
A list of summary statistics for the corresponding models
fitted with the calibrate_models function
}
\description{
This function is meant to be used after running the calibrate_models
function.
}

