% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sdsm-exploratory.R
\name{generate_table}
\alias{generate_table}
\title{Generate diagnostic plots for predictors and predictands.}
\usage{
generate_table(plotname, dataframe, y, conditional = FALSE,
  conditional_step = 1)
}
\arguments{
\item{plotname}{The name you want to save the plots as. IE:
if your data comes from the blogsville data-set, you might
want to call this "blogsville". If you want to save the plot
in a different directory, just specify this in this name:
"/home/blogsville" will}

\item{dataframe}{a dataframe object which contains a column
of dates, the predictand (response) variable, as well as the
predictor variables.}

\item{y}{The predictand (response) variable name from the
same dataframe.}

\item{conditional}{Whether you want the correlations to correspond
to the conditional model. If TRUE, then the function will return
correlations for the repsonse variable, which aspect of the conditional
model it will return is determined by the conditional_step parameter.}

\item{conditional_step}{Determines whether the first or second step of
the conditional model is returned. If 1, then this will return the
correlations with the 0/1 reponse variable. If 2, then the tables
will correspond to the orinal repsonse variable, but only on the
days in which the variable is a 1.}
}
\value{
A correlation matrix between predictor variables and
two PDF file diagnostic plots.
}
\description{
This function is meant to replicate the screen variables
section of the SDSM tool. It creates two plots, the table
which shows the correlations between predictors in all individual
months and annually, as well as the
}

